# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ChiefVoice CRM API
  version: 1.0.0
  description: |
    AI-powered voice CRM backend.

    **Authentication:** All protected endpoints require a identity-provider-issued ID token.
    Pass it as `Authorization: Bearer <token>`.
    The SSE endpoint `/api/logs-stream` uses a short-lived ticket created by `/api/logs-stream/ticket`
    because browser EventSource cannot set custom Authorization headers. Other
    protected endpoints do not accept JWTs in query strings.

    **Roles:** `Organization Admin`, `Super Admin`, `Agent`, `Manager`.
    Platform-admin endpoints additionally require the email to be in `PLATFORM_ADMIN_EMAILS`.

  contact:
    name: ChiefVoice Support
    email: support@chiefvoice.ai
servers:
  - url: /api
    description: Current server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication ID token
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
...
    CallLog:
      type: object
      properties:
        id:         { type: string }
        leadId:     { type: string }
        duration:   { type: integer }
        sentiment:  { type: string }
        transcript: { type: string }
        createdAt:  { type: string, format: date-time }
...
security:
  - BearerAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /calls/{id}/transcript:
    get:
      tags: [Calls]
      summary: Get transcript for a call
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Transcript }
        '404': { $ref: '#/components/responses/NotFound' }
  /calls/{id}/lead-responses:
    get:
      tags: [Calls]
      summary: Get lead qualification responses captured during a call
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Lead responses }
  /calls/{id}/analyze:
    post:
      tags: [Calls]
      summary: Re-run AI analysis on a call
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Analysis result }
  /call-logs:
    get:
      tags: [Calls]
      summary: List all call logs (outbound + inbound)
      responses:
        '200': { description: All call logs }
    post:
      tags: [Calls]
      summary: Create call log entry
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CallLog' }
      responses:
        '201': { description: Created }
  /call-logs/sync:
    post:
      tags: [Calls]
      summary: Bulk-upsert call logs
      requestBody:
        content:
          application/json:
            schema: { type: array, items: { type: object } }
      responses:
        '200': { description: Synced }
...
  /inbound-call-logs/sync:
    post:
      tags: [Calls]
      summary: Bulk-upsert inbound call logs
      requestBody:
        content:
          application/json:
            schema: { type: array, items: { type: object } }
      responses:
        '200': { description: Synced }

  # ── Campaigns ─────────────────────────────────────────────────────────────
...
  /vobiz/call:
    post:
      tags: [Telephony]
      summary: Initiate an outbound Vobiz call
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                to:     { type: string }
                leadId: { type: string }
      responses:
        '200': { description: Call initiated }
...